Add domain models & repo; refactor VM/manager
Introduce pure domain models and a repository-backed persistence layer, and refactor rendering and ViewModel to use them.

- Add DomainModels.kt: domain Page/Document/Stroke/Image/Pdf/Resource types with serialization and conversion helpers.
- Add DrawDocumentRepository.kt: repository to load full document trees (DocumentWithPages) and fast-save individual strokes to Room.
- Update DAOs (Daos.kt): add relational data classes (PageWithContent, DocumentWithPages), @Transaction queries, and convenience DAO methods.
- Refactor DrawManager: switch to domain models, map Ink strokes to domain strokes, update immediate-save flow to persist only new strokes, null-safety and small API cleanup (clip handling, enums compacted).
- Refactor DrawViewModel: become AndroidViewModel, use DrawDocumentRepository, load document asynchronously, expose saveNewStrokesToDatabase, and move UI document state into observable properties.
- Minor API fixes: Dimension calc* methods now accept Float resolution and DrawDocumentData adjusted to pass float values; various imports and logging additions.

These changes separate I/O from in-memory state, improve stroke persistence performance, and pave the way for clearer domain-driven rendering logic.

+++ app/src/main/java/com/studiomath/drawview/document/page/Dimension.kt
@@ -135,11 +135,11 @@ class Dimension(val width: Measure, val height: Measure) {
     }
 
     // TODO: 15/11/2021 Considerare la possibiltà di utilizzare i valori in px come Float
-    fun calcHeightFromResolutionPxInch(risoluzionePxInch: Int): Float {
+    fun calcHeightFromResolutionPxInch(risoluzionePxInch: Float): Float {
         return height.inch * risoluzionePxInch
     }
 
-    fun calcWidthFromResolutionPxInch(risoluzionePxInch: Int): Float {
+    fun calcWidthFromResolutionPxInch(risoluzionePxInch: Float): Float {
         return width.inch * risoluzionePxInch
     }
